refactor(Position): 패키지 분리 리팩터링
